pyproject.toml
[tool.poetry]
name = "novelist"
version = "0.1.0"
description = "CLI & web wizard for AI-assisted novel drafting"
authors = ["Bill Olson <bilols@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "novelist_cli" }]

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.16.0"
pydantic = "^2.7"
rich = "^14.0.0"
openai = "^1.88.0"
httpx = "^0.28.1"
tqdm = "^4.67.1"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.0"
mypy = "^1.16.1"
jsonschema = "^4.24.0"

[tool.poetry.scripts] 
novelist = "novelist_cli.__main__:app" 